cast

+++ io-multiplexing/select/main.c
@@ -68,7 +68,7 @@ int main(void)
         FD_ZERO(&readfds);
 
         // Add the server socket to the set
-        FD_SET(sockfd, &readfds);
+        FD_SET((unsigned int)sockfd, &readfds);
         max_fd = sockfd;
 
         // Add the client sockets to the set
@@ -78,7 +78,7 @@ int main(void)
 
             if(sd > 0)
             {
-                FD_SET(sd, &readfds);
+                FD_SET((unsigned int)sd, &readfds);
             }
 
             if(sd > max_fd)
@@ -97,7 +97,7 @@ int main(void)
         }
 
         // Handle new client connections
-        if(FD_ISSET(sockfd, &readfds))
+        if(FD_ISSET((unsigned int)sockfd, &readfds))
         {
             if((new_socket = accept(sockfd, (struct sockaddr *) &address, (socklen_t * ) & addrlen)) == -1)
             {
@@ -118,7 +118,7 @@ int main(void)
         {
             sd = client_sockets[i];
 
-            if(FD_ISSET(sd, &readfds))
+            if(FD_ISSET((unsigned int)sd, &readfds))
             {
                 char word_length;
                 char word[256];
@@ -131,7 +131,7 @@ int main(void)
                     // Connection closed or error
                     printf("Client %d disconnected\n", sd);
                     close(sd);
-                    FD_CLR(sd, &readfds); // Remove the closed socket from the set
+                    FD_CLR((unsigned int)sd, &readfds); // Remove the closed socket from the set
                     client_sockets[i] = 0;
                 }
                 else
@@ -143,7 +143,7 @@ int main(void)
                         // Connection closed or error
                         printf("Client %d disconnected\n", sd);
                         close(sd);
-                        FD_CLR(sd, &readfds); // Remove the closed socket from the set
+                        FD_CLR((unsigned int)sd, &readfds); // Remove the closed socket from the set
                         client_sockets[i] = 0;
                     }
                     else
